requirements.txt
flask
flask_sqlalchemy
python-dotenv
google-generativeai
pydantic-ai
celery
redis
